requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
pillow